data_collection/requirements.txt
praw==7.8.1
pandas==2.2.3
python-dotenv==1.1.0